Fix compiler warnings across the network library
- C4996: drop the premature [[deprecated]] on Request(const Request&, std::string);
  it is the mechanism make_request relies on to inherit session options and the
  advertised inherit_from() replacement was never implemented.
- C4244: cast the timeout seconds count() to long in set_timeout.
- C4189: mark the unused CURLSHcode as [[maybe_unused]] in set_share_option,
  keeping the TODO error-check intent.
- C4172: suppress a back-end false positive in WhenAll detail::get(); the returned
  reference points into the caller-owned GatherTasksStorage, not a local.
- C4101: drop the unused exception name in requestor_test's catch.

+++ src/Request.cpp
@@ -50,7 +50,7 @@ namespace asyncnet {
 	void Request::set_timeout(const std::optional<std::chrono::system_clock::duration>& timeout) {
 		using namespace std::chrono;
 
-		const auto timeout_secs = duration_cast<seconds>(timeout.value_or(seconds(0))).count();
+		const auto timeout_secs = static_cast<long>(duration_cast<seconds>(timeout.value_or(seconds(0))).count());
 		set_option<curlpp::options::Timeout>(timeout_secs);
 	}
